scx_mitosis: split out debug events
